+++ lib/src/rust/frb_generated.dart
@@ -3234,13 +3234,14 @@ class RustLibApiImpl extends RustLibApiImplPlatform implements RustLibApi {
   SendOptions dco_decode_send_options(dynamic raw) {
     // Codec=Dco (DartCObject based), see doc to use other codecs
     final arr = raw as List<dynamic>;
-    if (arr.length != 4)
-      throw Exception('unexpected arr length: expect 4 but see ${arr.length}');
+    if (arr.length != 5)
+      throw Exception('unexpected arr length: expect 5 but see ${arr.length}');
     return SendOptions(
       memo: dco_decode_opt_String(arr[0]),
       includeMemo: dco_decode_opt_box_autoadd_bool(arr[1]),
       pubkey: dco_decode_opt_String(arr[2]),
       metadata: dco_decode_opt_Map_String_String_None(arr[3]),
+      includeFee: dco_decode_opt_box_autoadd_bool(arr[4]),
     );
   }
 
@@ -4523,11 +4524,13 @@ class RustLibApiImpl extends RustLibApiImplPlatform implements RustLibApi {
     var var_includeMemo = sse_decode_opt_box_autoadd_bool(deserializer);
     var var_pubkey = sse_decode_opt_String(deserializer);
     var var_metadata = sse_decode_opt_Map_String_String_None(deserializer);
+    var var_includeFee = sse_decode_opt_box_autoadd_bool(deserializer);
     return SendOptions(
         memo: var_memo,
         includeMemo: var_includeMemo,
         pubkey: var_pubkey,
-        metadata: var_metadata);
+        metadata: var_metadata,
+        includeFee: var_includeFee);
   }
 
   @protected
@@ -5687,6 +5690,7 @@ class RustLibApiImpl extends RustLibApiImplPlatform implements RustLibApi {
     sse_encode_opt_box_autoadd_bool(self.includeMemo, serializer);
     sse_encode_opt_String(self.pubkey, serializer);
     sse_encode_opt_Map_String_String_None(self.metadata, serializer);
+    sse_encode_opt_box_autoadd_bool(self.includeFee, serializer);
   }
 
   @protected